requirements.txt
websockets==11.0.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-telegram-bot==20.4
schedule==1.2.0
python-dotenv==1.0.0
joblib==1.3.2